Fixed a crash in extensions::WebRequestEventDetails::SetFrameData()
The problem was that the code was following a PlzNavigate path were the
navigation_ui_data was supposed to be not null as it was dereferenced without
doing any null check. As properly stated
https://chromium-review.googlesource.com/c/chromium/src/+/757397 "with
PlzNavigate, there is always a NavigationUIData if there is a
ResourceRequestInfo (and ResourceRequestInfo is checked above in
web_request_api.cc)".

So the solution is to implement the required API that return the required
NavigationUIData object so that we don't crash.

https://crrev.com/1e5574e9aa99
https://crrev.com/059a6b4108cd